Make minibatch size an int.
Previously, when it was a float, it would give me a parsing error when
ParseInt() in C++ was called. I'm not sure why this hasn't been fixed yet.
Differences between python versions?

+++ egs/wsj/s5/steps/nnet3/train_dnn.py
@@ -134,7 +134,7 @@ def GetArgs():
                         help="If true, gpu is used with steps/nnet3/align.sh")
 
     # Parameters for the optimization
-    parser.add_argument("--trainer.optimization.minibatch-size", type=float, dest='minibatch_size',
+    parser.add_argument("--trainer.optimization.minibatch-size", type=int, dest='minibatch_size',
                         default = 512,
                         help="Size of the minibatch used to compute the gradient")
     parser.add_argument("--trainer.optimization.initial-effective-lrate", type=float, dest='initial_effective_lrate',
